breaking things up
